distributed-locks/ticketing/requirements.txt
flask==3.1.2
requests==2.32.5
psycopg2-binary==2.9.11